refactored all relevant files to non-testOnly directories
