+++ app/models/sessionData/SessionDataModel.scala
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package testOnly.models.sessionData
+package models.sessionData
 
 import play.api.libs.json.{Json, OFormat}
