+++ app/utils/SessionCookieUtil.scala
@@ -17,11 +17,11 @@
 package utils
 
 import config.AgentItvcErrorHandler
+import models.sessionData.SessionCookieData
+import models.sessionData.SessionDataPostResponse.{SessionDataPostFailure, SessionDataPostSuccess}
 import play.api.Logger
 import play.api.mvc.{Request, Result}
-import testOnly.models.sessionData.SessionCookieData
-import testOnly.models.sessionData.SessionDataPostResponse.{SessionDataPostFailure, SessionDataPostSuccess}
-import testOnly.services.SessionDataService
+import services.SessionDataService
 import uk.gov.hmrc.http.HeaderCarrier
 
 import scala.concurrent.{ExecutionContext, Future}
